T#774: Extract board (projects + tasks + comments + /api/board) to src/board/routes.ts
Phase 1.9 of Backend Modularization (Library #102). Mechanical move, no
logic changes.

**Moved to src/board/routes.ts (~584L)**:
- CREATE TABLE statements: projects, tasks, task_comments (idempotent IF NOT EXISTS)
- 5 projects routes: GET, POST, GET/:id, PATCH/:id, DELETE/:id
- 9 tasks routes: GET, POST, GET/:id, PATCH/:id, DELETE/:id,
  GET/:id/subtree, POST/bulk-status, GET/:id/comments, POST/:id/comments
- 1 board summary route: GET /api/board (kanban-grouped, project filter)

**Exports**:
- registerBoardRoutes(app, sqlite, { hasSessionAuth, requireBeastIdentity,
  isTrustedRequest, wsBroadcast }) — registers all 15 routes via DI

**Imports**:
- searchIndexUpsert, searchIndexDelete from ../search/routes.ts
  (tasks write handlers re-index for search)
- ../forum/mentions.ts dynamic-import for @-mention parsing in task
  comments (path corrected from ./forum/ to ../forum/)

**Server.ts changes**:
- Added import after daemons import
- Removed L5360-5919 block (~560 lines) — replaced with 4-line register call
- Net: server.ts down ~555 lines (was 10,483, now 9,928)

**Type safety**:
- Existing var was typed implicitly as {} — added 'as any' cast to match
  the rest of the codebase's sqlite query pattern
- Module pattern matches T#771/T#772/T#773 (closure-shadowed sqlite,
  generic-param TS2345 friction matches prior phases — T#788 sweep covers)

**Verification**:
- Boot test: server starts clean. All 15 board routes registered via DI.
- tsc --noEmit: zero new errors in src/board/routes.ts
- No logic changes: all auth gates preserved (PATCH /:id ownership check,
  POST /tasks Norm #79 in_review reviewer requirement, parent_task_id
  validation, status enum, etc.)
- /api/board kanban grouping by status with project filter preserved
- task_comments table CREATE preserved

**Cumulative across T#771-T#774**: server.ts -1593 lines (was 11,963, now 9,928)
